REFACTOR: Category components
Addresses: #90
