+++ .idea/runConfigurations/_template__of_Pest.xml
@@ -8,6 +8,9 @@
     <option name="pestRunnerSettings">
       <PestRunner method="" />
     </option>
+    <option name="runnerSettings">
+      <PhpTestRunnerSettings method="" />
+    </option>
     <method v="2" />
   </configuration>
 </component>
