+++ .idea/runConfigurations/_template__of_Pest2.xml
@@ -1,0 +1,16 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="true" type="PestRunConfigurationType">
+    <CommandLine workingDirectory="$PROJECT_DIR$/app">
+      <envs>
+        <env name="DB_DATABASE" value="laravel_testing" />
+      </envs>
+    </CommandLine>
+    <option name="pestRunnerSettings">
+      <PestRunner method="" />
+    </option>
+    <option name="runnerSettings">
+      <PhpTestRunnerSettings method="" />
+    </option>
+    <method v="2" />
+  </configuration>
+</component>
