+++ app/app/Livewire/Dto/GameboardInformationForCategory.php
@@ -10,6 +10,7 @@ use Domain\Definitions\Konjunkturphase\ValueObject\CategoryId;
 class GameboardInformationForCategory
 {
     /**
+     * @param string $componentName
      * @param CategoryId $title
      * @param int|null $kompetenzen
      * @param int|null $kompetenzenRequiredByPhase
@@ -18,6 +19,7 @@ class GameboardInformationForCategory
      * @param PileId|null $cardPile
      */
     public function __construct(
+        public string $componentName,
         public CategoryId $title,
         public ?int $kompetenzen,
         public ?int $kompetenzenRequiredByPhase,
