+++ app/resources/views/components/gameboard/categories/categories-bildung.blade.php
@@ -1,0 +1,7 @@
+@props([
+    'category' => null,
+    'gameEvents' => null,
+])
+
+<x-card-pile :category="$category->title->value" :card-pile="$category->cardPile->value"
+     :game-events="$gameEvents"/>
