+++ app/resources/views/components/gameboard/categories/categories-investitionen.blade.php
@@ -1,0 +1,4 @@
+@props([
+    'category' => null,
+    'gameEvents' => null,
+])
